smp: Use uintptr_t on apboot pointers
Make smp 64b compatible.

Message-ID: <20260107030711.582699-7-damien@zamaudio.com>

+++ i386/i386/mp_desc.c
@@ -285,7 +285,7 @@ start_other_cpus(void)
 
 	//Copy cpu initialization assembly routine
 	memcpy((void*) phystokv(apboot_addr), (void*) &apboot,
-	       (uint32_t)&apbootend - (uint32_t)&apboot);
+	       (uintptr_t)&apbootend - (uintptr_t)&apboot);
 
 	unsigned cpu = cpu_number_slow();
